Organize API modules by API version and resource type
