+++ pyinaturalist/v2/__init__.py
@@ -1,0 +1,1 @@
+"""Coming soon!"""
