+++ test/v0/test_observations_v0.py
@@ -8,14 +8,12 @@ from requests import HTTPError
 
 from pyinaturalist.constants import API_V0_BASE_URL
 from pyinaturalist.exceptions import ObservationNotFound
-from pyinaturalist.rest_api import (
-    OBSERVATION_FORMATS,
+from pyinaturalist.request_params import OBSERVATION_FORMATS
+from pyinaturalist.v0 import (
     add_photo_to_observation,
     create_observation,
     delete_observation,
-    get_observation_fields,
     get_observations,
-    put_observation_field_values,
     update_observation,
 )
 from test.conftest import load_sample_data
@@ -56,73 +54,6 @@ def test_get_observations__invalid_format(response_format):
         get_observations(taxon_id=493595, response_format=response_format)
 
 
-def test_get_observation_fields(requests_mock):
-    """get_observation_fields() work as expected (basic use)"""
-    page_2_json_response = load_sample_data('get_observation_fields_page2.json')
-
-    requests_mock.get(
-        f'{API_V0_BASE_URL}/observation_fields.json?q=sex&page=2',
-        json=page_2_json_response,
-        status_code=200,
-    )
-
-    response = get_observation_fields(q='sex', page=2)
-    first_result = response['results'][0]
-    assert len(response['results']) == 3
-    assert first_result['id'] == 5
-    assert first_result['datatype'] == 'text'
-    assert first_result['created_at'] == datetime(2012, 1, 23, 8, 12, 0, 138000, tzinfo=tzutc())
-    assert first_result['updated_at'] == datetime(2018, 10, 16, 6, 47, 43, 975000, tzinfo=tzutc())
-
-
-def test_get_observation_fields__all_pages(requests_mock):
-    page_1_json_response = load_sample_data('get_observation_fields_page1.json')
-    page_2_json_response = load_sample_data('get_observation_fields_page2.json')
-
-    requests_mock.get(
-        f'{API_V0_BASE_URL}/observation_fields.json?q=sex&page=1',
-        json=page_1_json_response,
-        status_code=200,
-    )
-    requests_mock.get(
-        f'{API_V0_BASE_URL}/observation_fields.json?q=sex&page=2',
-        json=page_2_json_response,
-        status_code=200,
-    )
-    requests_mock.get(
-        f'{API_V0_BASE_URL}/observation_fields.json?q=sex&page=3',
-        json=[],
-        status_code=200,
-    )
-    expected_results = page_1_json_response + page_2_json_response
-
-    response = get_observation_fields(q='sex', page='all')
-    first_result = response['results'][0]
-    assert response['total_results'] == len(response['results']) == len(expected_results)
-    assert first_result['created_at'] == datetime(2016, 5, 29, 16, 17, 8, 51000, tzinfo=tzutc())
-    assert first_result['updated_at'] == datetime(2018, 1, 1, 1, 17, 56, 7000, tzinfo=tzutc())
-
-
-def test_put_observation_field_values(requests_mock):
-    requests_mock.put(
-        f'{API_V0_BASE_URL}/observation_field_values/31',
-        json=load_sample_data('put_observation_field_value_result.json'),
-        status_code=200,
-    )
-
-    response = put_observation_field_values(
-        observation_id=18166477,
-        observation_field_id=31,  # Animal behavior
-        value='fouraging',
-        access_token='valid token',
-    )
-
-    assert response['id'] == 31
-    assert response['observation_field_id'] == 31
-    assert response['observation_id'] == 18166477
-    assert response['value'] == 'fouraging'
-
-
 def test_update_observation(requests_mock):
     requests_mock.put(
         f'{API_V0_BASE_URL}/observations/17932425.json',
@@ -142,8 +73,8 @@ def test_update_observation(requests_mock):
     assert response[0]['description'] == 'updated description v2 !'
 
 
-@patch('pyinaturalist.rest_api.ensure_file_objs')
-@patch('pyinaturalist.rest_api.put')
+@patch('pyinaturalist.v0.observations.ensure_file_objs')
+@patch('pyinaturalist.v0.observations.put')
 def test_update_observation__local_photo(put, ensure_file_objs):
     update_observation(1234, access_token='token', local_photos='photo.jpg')
 
@@ -205,8 +136,8 @@ def test_create_observation(requests_mock):
     assert response[0]['taxon_id'] == 55626  # Pieris Rapae @ iNaturalist
 
 
-@patch('pyinaturalist.rest_api.ensure_file_objs')
-@patch('pyinaturalist.rest_api.post')
+@patch('pyinaturalist.v0.observations.ensure_file_objs')
+@patch('pyinaturalist.v0.observations.post')
 def test_create_observation__local_photo(post, ensure_file_objs):
     create_observation(access_token='token', local_photos='photo.jpg')
